+++ tidy3d/web/core/environment.py
@@ -3,6 +3,7 @@ import os
 from .core_config import get_logger
 
 from pydantic.v1 import BaseSettings, Field
+import ssl
 
 
 class EnvironmentConfig(BaseSettings):
@@ -17,6 +18,7 @@ class EnvironmentConfig(BaseSettings):
     s3_region: str
     ssl_verify: bool = Field(True, env="TIDY3D_SSL_VERIFY")
     enable_caching: bool = True
+    ssl_version: ssl.TLSVersion = None
 
     def active(self) -> None:
         """Activate the environment instance."""
@@ -175,6 +177,16 @@ class Environment:
             If `False`, do not duplicate checking. Just run the task directly.
         """
         self._current.enable_caching = enable_caching
+
+    def set_ssl_version(self, ssl_version: ssl.TLSVersion) -> None:
+        """Set the ssl version.
+
+        Parameters
+        ----------
+        ssl_version : ssl.TLSVersion
+            The ssl version to set.
+        """
+        self._current.ssl_version = ssl_version
 
 
 Env = Environment()